Finishing up with the linux stuff
So I moved the project to linux and I added new stuff, but at the same
time fixed my cmake and other things.

Right now there's nothing that installs the project dependencies, I will
add that later so that you're able to build it directly with a bash file
instead of cmake.